ZK sec audit 3 (#125)
* Issue number 1

Typo in Field Chip causing inv(0) = 0.

* Issue number 3

Variable-length poseidon miscomputes hashes for short inputs

* Issue number 5

Verifier panics on empty committed instances

* Issue number 6

Integer overflow in blinding factors calculation.

I decided to panic instead of having a result.

* Issue number 0

Add all the limbs of y to the PI representation of a point.

* Issue number 7

Unify logic on committed instance queries.
Align the prover logic with the one of the verifier.

* Update CHANGELOG files

---------